GPUDevice is super fast
